Refactor null-checking and price history update logic
Removed unnecessary null-checking logic on Grocery Item properties. This change was done because it was throwing unnecessary exceptions. Also, refactored price history update segment on the UpdateGroceryItem() method for readability and coherence. It extracts a common sequence of code into a new method: UpdatePriceHistory(). This will improve Maintainability of the code.